dont break ABI
There is a final, important issue with cpu/interrupt.asm. The common stubs create an instance of struct registers on the stack and then call the C handler. But that breaks the ABI, since the stack belongs to the called function and they may change them as they please. It is needed to pass the struct as a pointer.

To achieve this, edit cpu/isr.h and cpu/isr.c and change registers_t r into registers_t *t, then, instead of accessing the fields of the struct via ., access the fields of the pointer via ->. Finally, in cpu/interrupt.asm, and add a push esp before calling both isr_handler and irq_handler -- remember to also pop eax to clear the pointer afterwards.